# Improve performance by making state watchers periodic