--- a/deps.edn
+++ b/deps.edn
@@ -14,6 +14,7 @@
   com.evocomputing/colors {:mvn/version "1.0.6"}
   com.fzakaria/slf4j-timbre {:mvn/version "0.3.21"}
   com.taoensso/timbre {:mvn/version "5.1.2"}
+  com.taoensso/tufte {:mvn/version "2.2.0"}
   commons-io/commons-io {:mvn/version "2.8.0"}
   crouton/crouton {:mvn/version "0.1.2"}
   crypto-random/crypto-random {:mvn/version "1.2.1"}