Before
```
:tufte/nil-id,
pId                                       nCalls        Min      50% ≤      90% ≤      95% ≤      99% ≤        Max       Mean   MAD      Clock  Total

:battle-mod-details-watcher                1,082     2.30μs     2.69ms     4.34ms     6.58ms    12.42ms    43.10ms     2.68ms  ±59%     2.90s     53%
:battle-map-details-watcher                1,082     2.10μs     1.09ms     1.89ms     2.47ms     6.79ms    32.11ms     1.25ms  ±54%     1.35s     25%
:update-battle-status-sync-watcher         1,082     5.80μs    13.40μs     3.24ms     6.20ms     8.12ms    15.93ms   693.09μs ±172%   749.92ms    14%
:spring-isolation-dir-changed-watcher      1,082     0.00ns   320.60μs   514.90μs   677.70μs     1.76ms     8.75ms   356.73μs  ±46%   385.98ms     7%
:auto-get-resources-watcher                1,082     4.90μs    26.40μs    47.30μs    57.50μs   222.00μs     7.71ms    58.47μs ±101%    63.26ms     1%
:fix-spring-isolation-dir-watcher          1,082     0.00ns     7.20μs    13.50μs    17.20μs    43.40μs   377.10μs     8.44μs  ±54%     9.13ms     0%
:replay-map-and-mod-details-watcher        1,082     2.70μs     4.70μs     7.90μs    13.90μs    54.10μs   878.40μs     8.35μs  ±77%     9.04ms     0%
:fix-selected-server-watcher               1,082     1.90μs     4.30μs     8.40μs    11.20μs    42.40μs   268.80μs     6.51μs  ±62%     7.05ms     0%
:fix-missing-resource-watcher              1,082     2.10μs     4.00μs     6.30μs    11.20μs    60.50μs   368.10μs     6.30μs  ±64%     6.82ms     0%

Accounted                                                                                                                               5.48s    100%
Clock                                                                                                                                   5.50s    100%

```

After
```
:spring-lobby/chimer,
pId                                       nCalls        Min      50% ≤      90% ≤      95% ≤      99% ≤        Max       Mean   MAD      Clock  Total

:update-battle-status-sync-watcher            60    17.10μs    32.20μs     7.95ms     9.53ms    28.50ms    28.65ms     1.99ms ±166%   119.54ms    89%
:battle-mod-details-watcher                   60    18.40μs    37.30μs     3.58ms     3.74ms     4.97ms    25.43ms   990.13μs ±160%    59.41ms    44%
:battle-map-details-watcher                   60    16.80μs    37.20μs     1.71ms     2.09ms     2.91ms    25.96ms   765.44μs ±154%    45.93ms    34%

Accounted                                                                                                                             224.87ms   168%
Clock                                                                                                                                 134.21ms   100%


:skylobby/ui,
pId                                       nCalls        Min      50% ≤      90% ≤      95% ≤      99% ≤        Max       Mean   MAD      Clock  Total

:battle-view                                  10     2.82ms     3.08ms     3.20ms     3.34ms     3.34ms     3.34ms     3.04ms   ±4%    30.41ms    58%
:root                                         40   450.10μs   528.00μs   618.70μs   671.20μs   702.90μs   702.90μs   535.54μs   ±8%    21.42ms    41%
:replays-window                                9     5.00μs     6.30μs     7.80μs     8.80μs     8.80μs     8.80μs     6.65μs  ±13%    59.81μs     0%

Accounted                                                                                                                              51.89ms    99%
Clock                                                                                                                                  52.31ms   100%


:spring-lobby/state-watcher,
pId                                       nCalls        Min      50% ≤      90% ≤      95% ≤      99% ≤        Max       Mean   MAD      Clock  Total

:spring-isolation-dir-changed-watcher        224    99.00ns   201.00ns   387.40μs   451.50μs   665.00μs     1.12ms    85.46μs ±159%    19.14ms    44%
:auto-get-resources-watcher                  224     6.10μs    17.90μs    51.20μs    77.80μs   221.40μs   591.70μs    28.92μs  ±69%     6.48ms    15%
:replay-map-and-mod-details-watcher          224     4.40μs     8.50μs    16.40μs    27.60μs   129.30μs     1.62ms    18.88μs  ±99%     4.23ms    10%
:fix-missing-resource-watcher                224     2.80μs     5.60μs    15.70μs    25.60μs   128.20μs   327.00μs    11.63μs  ±88%     2.60ms     6%
:fix-selected-server-watcher                 224     2.40μs     6.70μs    14.10μs    19.80μs    62.80μs   100.50μs     9.65μs  ±58%     2.16ms     5%
:fix-spring-isolation-dir-watcher            224    99.00ns   300.00ns     9.50μs    11.70μs    19.30μs    46.20μs     2.59μs ±137%   580.31μs     1%

Accounted                                                                                                                              35.19ms    81%
Clock                                                                                                                                  43.60ms   100%

```